Many changes to the App
A method was found for adding a map grid. This method was made into a subclass of a JPanel. The login window recieved one of these, along with sliders, which allows picking of coordinates. The shortest route to a coordinate is able to be found from various locations, and lines, as well as circles, can be drawn on the grid. The grid itself can change the number of cells it contains. The grid is also capable of many other methods.

Other various minor changes were made, but they won't be detailed here.